ショットガン使用時の命中数・命中率計算を改善、死体として静止するまで銃弾・手榴弾・ゾンビ攻撃の当たり判定を行うように。
git-svn-id: svn://svn.pf.osdn.net/x/xo/xops-mikan/OpenXOPS/trunk@187 ff51d050-6dfb-4b14-850f-860e8d883e2a

+++ gamemain.h
@@ -82,7 +82,7 @@ struct GameInfo{
 	bool missioncomplete;	//!< ƒ~ƒbƒVƒ‡ƒ“Š®—¹
 	unsigned int framecnt;	//!< ƒtƒŒ[ƒ€”
 	int fire;		//!< ŽËŒ‚‰ñ”
-	int ontarget;	//!< –½’†”
+	float ontarget;	//!< –½’†”
 	int kill;		//!< “|‚µ‚½“G‚Ì”	
 	int headshot;	//!< “G‚Ì“ª•”‚É–½’†‚µ‚½”
 };
